feat: allow fractional quantities (decimal) on order items
- OrderItem.Quantity: int -> decimal
- CreateOrderItemDto, OrderItemResponseDto: int -> decimal
- Migration: ALTER COLUMN Quantity TYPE numeric(18,2)
- Pedidos.razor: MudNumericField decimal, display N2, receipt HTML N2
- ReceiptFormatter: {Quantity:N2} format

+++ ScaleAtacado.Application/DTOs/CreateOrderItemDto.cs
@@ -1,3 +1,3 @@
 namespace ScaleAtacado.Application.DTOs;
 
-public record CreateOrderItemDto(Guid ProductId, int Quantity);
+public record CreateOrderItemDto(Guid ProductId, decimal Quantity);
